Download file of Version1.(Android only)
In this version, we designed a simple frame which contains self information page, rule page, music setting page, "playing alone" pattern, "create new room to play" pattern and "join existing room to play" pattern. But we only achieved playing alone pattern and joining an existing three-persons room to play with other two friends in this version.
This is the download file of Version1. (Only for Android system)